requirements.txt
# python3.6
h5py
# nni
numpy
opencv_python
Pillow
scipy
# torch==1.8.0+cu111
# torchvision==0.9.0+cu111
tqdm
